requirements.txt
numpy==1.26.4
accelerate==0.31.0
altair==4.2.2
contourpy==1.2.1
datasets==2.20.0
evaluate==0.4.2
faiss-cpu==1.8.0
keras==3.3.3
lamini==2.2.1
langchain==0.2.1
pandas==2.2.1
pytorch-pretrained-bert==0.6.2
scipy==1.13.0
seqeval==1.2.2
tensorboard==2.16.2
tensorflow==2.16.1
torchvision==0.17.2
transformers==4.42.3


